gensitemap/requirements.txt
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
pip-chill==1.0.3
pylint-django==2.5.5
pytest==8.2.2
scrapy==2.11.0 ; python_version >= "3.10" and python_version < "4.0"
scrapyd
scrapyd-client
